requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
google-genai>=1.0.0
dnspython>=2.6.0

